% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_createPrevalence.R
\name{createPrevalence}
\alias{createPrevalence}
\title{Create Prevalence Data Frames from a Topic Model}
\usage{
createPrevalence(model, sort = FALSE)
}
\arguments{
\item{model}{A topic model object.}

\item{sort}{A logical value indicating whether to sort the resulting data frames by document. Default is FALSE.}
}
\value{
A list containing three data frames:
\item{td_gamma}{A data frame with the document-topic matrix.}
\item{gamma_dummies}{A wide-format data frame with dummy variables for each topic.}
\item{top_gamma}{A data frame with the top topic for each document and its corresponding gamma value.}
}
\description{
This function generates three data frames from a given topic model:
a document-topic matrix, a wide-format document-topic matrix with dummy variables,
and a data frame with the top topic for each document.
}
\examples{
\dontrun{
model <- stm::stm(documents, vocab, K = 5, max.em.its = 75)
prevalence_data <- createPrevalence(model, sort = TRUE)
}

}
