% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osf_file.R
\name{download_osf_file}
\alias{download_osf_file}
\title{Download OSF File}
\usage{
download_osf_file(
  local_folder,
  osf_file_id = NULL,
  local_directory = "local_data/osf_pull"
)
}
\arguments{
\item{local_folder}{A character vector specifying the local folder(s) where the data should be downloaded.}

\item{osf_file_id}{A character vector specifying the OSF file ID(s). Default is NULL.}

\item{local_directory}{A character string specifying the local directory where the data should be stored. Default is "local_data/osf_pull".}
}
\value{
None. The function is called for its side effects (downloading data).
If an error occurs during the download of a file, a message will be printed with the file ID and the error message.
}
\description{
This function downloads data from the Open Science Framework (OSF) by specifying a file ID.
}
\details{
You must specify \code{osf_file_id}. The function will create the necessary local directories if they do not exist.
}
\examples{
\dontrun{
# Download data from an OSF file
download_osf_file(local_folder = c("folder1", "folder2"), osf_file_id = c("file_id1", "file_id2"))
}

}
