% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushover_status.R
\name{pushover_status}
\alias{pushover_status}
\title{Send a Pushover Notification for Job Status}
\usage{
pushover_status(pushover_title, pushover_job, status)
}
\arguments{
\item{pushover_title}{A character string specifying the title of the notification.}

\item{pushover_job}{A character string specifying the name of the job.}

\item{status}{A character string specifying the status of the job.
Must be either "started" or "finished".}
}
\value{
Sends a Pushover notification and returns the response object.
}
\description{
This function sends a Pushover notification indicating the status of a job.
}
\examples{
\dontrun{
pushover_status("Job Notification", "Data Analysis", "started")
pushover_status("Job Notification", "Data Analysis", "finished")
}
}
