% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osf_component.R
\name{download_osf_component}
\alias{download_osf_component}
\title{Download data folder from OSF}
\usage{
download_osf_component(
  local_folder,
  osf_component_id,
  local_directory = "local_data/osf_pull"
)
}
\arguments{
\item{local_folder}{A character vector specifying the folder(s) to save the downloaded data into. Will be used to add to argument 'local_directory'.}

\item{osf_component_id}{A character vector specifying the corresponding node ID(s) for each component.}

\item{local_directory}{The directory where data will be downloaded. Default is "local_data/osf_pull".}
}
\value{
No explicit return value. Data will be downloaded to the specified directory.
}
\description{
This function is design to simplify and help download files from OSF, especially from multiple OSF components (based on their ID) at once.
}
\examples{
\dontrun{
local_folder <- c("base")
osf_component <- c("jmipd")
download_osf_data(local_folder, osf_component)
}
}
