% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushover_status.R
\name{pushover_status}
\alias{pushover_status}
\title{Send a Pushover Notification for Job Status}
\usage{
pushover_status(pushover_title = NULL, pushover_job, status)
}
\arguments{
\item{pushover_title}{A character string specifying the title of the notification. Defaults to NULL.}

\item{pushover_job}{A character string specifying the name of the job.}

\item{status}{A character string specifying the status of the job.
Must be either "started" or "finished".}
}
\value{
Sends a Pushover notification and returns the response object.
}
\description{
This function sends a Pushover notification indicating the status of a job.
It relies heavily on the \code{pushoverr::pushover()} function and requires
the same arguments to be defined to run properly.
}
\details{
This function uses the \code{pushoverr::pushover()} function to send
the notification. Ensure you have the \code{pushoverr} package installed and
configured correctly with your Pushover user key and application token.
}
\examples{
\dontrun{
pushover_status("Job Notification", "Data Analysis", "started")
pushover_status("Job Notification", "Data Analysis", "finished")
pushover_status(NULL, "Data Analysis", "started")
}
}
